core: GrallocOps: fix vector type
1. gralloc 0.3 does not support vector<uint64_t>, so it is changed back
to vector<int>.
2. Remove support for drm_fourcc/modifier in normal API.

fixup: d914e44f57d6 (im2d_api: supports auto-config store_mode by GraphicBuffer)

update to 1.10.5_[3]

+++ include/GrallocOps.h
@@ -48,7 +48,7 @@
 #include <cutils/native_handle.h>
 #include "rga.h"
 
-typedef std::vector<uint64_t> rga_gralloc_attr_t;
+typedef std::vector<int> rga_gralloc_attr_t;
 
 // -------------------------------------------------------------------------------
 int         RkRgaGetHandleFd(buffer_handle_t handle, int *fd);
